Change arch to arm64 and libqwunits packging works
Had to change from aarch64 to arm64. THis had to do with checking for
the correct architecture for the libc++ and fmt9 dependencies.

Also, got the libqwunits packge to install and uninstall correctly.
The liqwunits-dev complains about garbage dependencies in the package.
It has a line with acomma in it. I am trying to set the libqwunits-dev
to be dependent on the runtime component. But maybe I have to set it
as a specific dependency instead of tryng to say the dev component
depends on the runtime component.